+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -125,7 +125,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2026-02-05T13:10:37.747Z",
+    "2026-02-05T14:12:56.782Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {