+++ packages/orq-rc/RELEASES.md
@@ -6449,3 +6449,13 @@ Based on:
 - [typescript v4.3.0-rc.29] packages/orq-rc
 ### Releases
 - [NPM v4.3.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.29 - packages/orq-rc
+
+## 2026-02-05 14:15:49
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.705.0 (2.808.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.3.0-rc.30] packages/orq-rc
+### Releases
+- [NPM v4.3.0-rc.30] https://www.npmjs.com/package/@orq-ai/node/v/4.3.0-rc.30 - packages/orq-rc